+++ spec/integration/braintree/transaction_spec.rb
@@ -2116,6 +2116,61 @@ describe Braintree::Transaction do
         expect(result.transaction.paypal_details.debug_id).not_to be_nil
       end
 
+      it "can create a transaction with a fake meta checkout card nonce" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :payment_method_nonce => Braintree::Test::Nonce::MetaCheckoutCard,
+        )
+
+        result.success?.should == true
+        result.transaction.should_not be_nil
+        meta_checkout_card_details = result.transaction.meta_checkout_card_details
+        meta_checkout_card_details.should_not be_nil
+        meta_checkout_card_details.bin.should == "401288"
+        meta_checkout_card_details.card_type.should == Braintree::CreditCard::CardType::Visa
+        meta_checkout_card_details.cardholder_name.should == "Meta Checkout Card Cardholder"
+        meta_checkout_card_details.container_id.should == "container123"
+        meta_checkout_card_details.customer_location.should == "US"
+        meta_checkout_card_details.expiration_date.should == "12/2024"
+        meta_checkout_card_details.expiration_year.should == "2024"
+        meta_checkout_card_details.expiration_month.should == "12"
+        meta_checkout_card_details.image_url.should == "https://assets.braintreegateway.com/payment_method_logo/visa.png?environment=development"
+        meta_checkout_card_details.is_network_tokenized.should == false
+        meta_checkout_card_details.last_4.should == "1881"
+        meta_checkout_card_details.masked_number.should == "401288******1881"
+        meta_checkout_card_details.prepaid.should == "No"
+      end
+
+      it "can create a transaction with a fake meta checkout token nonce" do
+        result = Braintree::Transaction.create(
+          :type => "sale",
+          :amount => Braintree::Test::TransactionAmounts::Authorize,
+          :payment_method_nonce => Braintree::Test::Nonce::MetaCheckoutToken,
+        )
+
+        result.success?.should == true
+        result.transaction.should_not be_nil
+        meta_checkout_token_details = result.transaction.meta_checkout_token_details
+
+        meta_checkout_token_details.should_not be_nil
+        meta_checkout_token_details.bin.should == "401288"
+        meta_checkout_token_details.card_type.should == Braintree::CreditCard::CardType::Visa
+        meta_checkout_token_details.cardholder_name.should == "Meta Checkout Token Cardholder"
+        meta_checkout_token_details.container_id.should == "container123"
+        meta_checkout_token_details.cryptogram.should == "AlhlvxmN2ZKuAAESNFZ4GoABFA=="
+        meta_checkout_token_details.customer_location.should == "US"
+        meta_checkout_token_details.ecommerce_indicator.should == "07"
+        meta_checkout_token_details.expiration_date.should == "12/2024"
+        meta_checkout_token_details.expiration_year.should == "2024"
+        meta_checkout_token_details.expiration_month.should == "12"
+        meta_checkout_token_details.image_url.should == "https://assets.braintreegateway.com/payment_method_logo/visa.png?environment=development"
+        meta_checkout_token_details.is_network_tokenized.should == true
+        meta_checkout_token_details.last_4.should == "1881"
+        meta_checkout_token_details.masked_number.should == "401288******1881"
+        meta_checkout_token_details.prepaid.should == "No"
+      end
+
       it "can create a transaction with a fake apple pay nonce" do
         customer = Braintree::Customer.create!
         result = Braintree::Transaction.create(